chatbot/requirements.txt
fastapi
fastapi-sessions
python-multipart
uvicorn[standard]
tortoise-orm[aiomysql]
sentence-transformers
chromadb
openai
langchain
langchain-community
langchain_openai
langchain-chroma
opencv-python==4.11.0.86
tensorflow==2.19.0
Pillow==11.0.0
mtcnn==1.0.0
#dlib
pytest
pytest-asyncio